feat(ui): Implement tab-based navigation and refactor pages
