+++ frontend/src/components/Navigation.tsx
@@ -1,0 +1,12 @@
+// src/components/Navigation.tsx
+
+import { Link } from 'react-router-dom';
+
+export function Navigation() {
+  return (
+    <nav style={{ marginBottom: '2rem', borderBottom: '1px solid #555', paddingBottom: '1rem' }}>
+      <Link to="/" style={{ marginRight: '1rem', fontSize: '1.2em' }}>コード分析</Link>
+      <Link to="/scale-analyzer" style={{ fontSize: '1.2em' }}>スケール分析</Link>
+    </nav>
+  );
+}
